pyproject.toml
[project]
name = "elementary-math-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "hauuto",email = "thanhhau670@gmail.com"}
]
requires-python = "^3.10,<3.15"
dependencies = [
    "torch (>=2.11.0,<3.0.0)",
    "transformers (>=5.7.0,<6.0.0)",
    "peft (>=0.19.1,<0.20.0)",
    "bitsandbytes (>=0.49.2,<0.50.0)",
    "paddleocr (>=3.5.0,<4.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn[standard] (>=0.46.0,<0.47.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.2.0,<2.0.0)",
    "gradio (>=6.13.0,<7.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "nltk (>=3.9.4,<4.0.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "datasets (>=4.8.5,<5.0.0)",
    "trl (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "paddlepaddle (>=3.3.1,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-genai (>=1.74.0,<2.0.0)",
    "pandas (<2.3.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

